Fix "invisible user switch applet" problem
There is a period of time when first starting up that
get_user calls will fail to function properly because
it communicates with the accounts service synchronously
while the user manager is doing it's initial bring up
asynchronously.

This package makes get_user asynchronous like the rest of
the user manager code, and also enforces an ordering
requirement that get_user requests are deferred until
after the user manager is loaded.

The upshot is that the user switcher should consistently
present the user's name now when the accounts service
is disabled.